Updated SRA Toolkit to 2.11.0
This was an attempt to solve PTV-1678 where SRA objects loaded from NCBI are
not able to be merged. This fix did not solv the problem, but is enough of a
chore to leave this here for anyone with a need to do so in the future.

The SRA toolkit now requires being configured before fastq-dump will run
successfully. The only addition needed to the default configuration for this to
work is setting a value in /LIBS/GUID. The options are specified via these
slash delimited strings. If you need to change a specific option then first run
vdb-config -n to print out the options and values to get the correct key.